+++ Changelog.md
@@ -7,3 +7,4 @@
 * Show more version info in logs
 * Do not update `SharedFunctional` repo
 * Use default branch for `OnlineDocuments` repo
+* Enable autostart of `ds:exmaple`